video-agent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-openai==0.0.5
openai-whisper==20231117
openai==1.10.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
